polished up components
